LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/drawing-rectangles-to-the-canvas/2d.clearRect.clip.worker-expected.txt


PASS clearRect is affected by clipping regions 
